+++ Cargo.toml
@@ -222,7 +222,7 @@ burn-wgpu = { path = "crates/burn-wgpu", version = "0.22.0-pre.1", default-featu
 cubecl = { git = "https://github.com/tracel-ai/cubecl", default-features = false, rev = "901d37dbd4a916b764360e4fea111b01356f8263" }
 cubecl-common = { git = "https://github.com/tracel-ai/cubecl", default-features = false, rev = "901d37dbd4a916b764360e4fea111b01356f8263" }
 cubecl-zspace = { git = "https://github.com/tracel-ai/cubecl", default-features = false, rev = "901d37dbd4a916b764360e4fea111b01356f8263" }
-cubek = { git = "https://github.com/tracel-ai/cubek", default-features = false, rev = "9372f94bfa8f66c6843bb9c37bd30c6fbcdb45cf" }
+cubek = { git = "https://github.com/tracel-ai/cubek", default-features = false, rev = "41b8c27d92c6ed5dd4e5b9fe77a9ed8665025562" }
 ### For local development. ###
 # cubecl = { path = "../cubecl/crates/cubecl", default-features = false }
 # cubecl-common = { path = "../cubecl/crates/cubecl-common", default-features = false }
