+++ crates/burn-cubecl/src/kernel/pool/mod.rs
@@ -1,15 +1,3 @@
-mod adaptive_avg_pool2d;
-mod adaptive_avg_pool2d_backward;
-mod avg_pool2d;
-mod avg_pool2d_backward;
-mod max_pool2d;
-mod max_pool2d_backward;
+mod base;
 
-pub(super) mod pool2d;
-
-pub(crate) use adaptive_avg_pool2d::*;
-pub(crate) use adaptive_avg_pool2d_backward::*;
-pub(crate) use avg_pool2d::*;
-pub(crate) use avg_pool2d_backward::*;
-pub(crate) use max_pool2d::*;
-pub(crate) use max_pool2d_backward::*;
+pub(crate) use base::*;
